Fixed minor warning detected in Visual Studio.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3260 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/emucore/Paddles.cxx
@@ -411,7 +411,7 @@ void Paddles::setMouseSensitivity(int sensitivity)
 void Paddles::setPaddleRange(int range)
 {
   range = BSPF_clamp(range, 1, 100);
-  TRIGRANGE = TRIGMAX * (range / 100.0);
+  TRIGRANGE = int(TRIGMAX * (range / 100.0));
 }
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
